+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.10.0 - 
+
+### Changed
+- Use `<small>` tag for posts info wrapper
+
 ## 0.9.2 - 2018-07-16
 
 ### Changed